requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
mysql-connector-python
python-dotenv
email-validator
